+++ lib/generators/selections_scaffold/templates/selection.rb
@@ -1,0 +1,3 @@
+class Selection < ActiveRecord::Base
+  selectable
+end