+++ lib/selections/selectable.rb
@@ -13,17 +13,17 @@ module Selections
         acts_as_tree
 
         validates_presence_of :name
-        validates_uniqueness_of :name, :scope => :parent_id
-        validates_uniqueness_of :system_code, :scope => :archived_at
-        validates_format_of :system_code, :with => /^[a-z][a-zA-Z_0-9]*$/, :message => "can only contain alphanumeric characters and '_', not spaces"
+        validates_uniqueness_of :name, scope: :parent_id
+        validates_uniqueness_of :system_code, scope: :archived_at
+        validates_format_of :system_code, :with => /\A[a-z][a-zA-Z_0-9]*\Z/, :message => "can only contain alphanumeric characters and '_', not spaces"
 
-        before_validation :auto_gen_system_code, :on => :create
-        before_validation :disable_system_code_change, :on => :update
+        before_validation :auto_gen_system_code, on: :create
+        before_validation :disable_system_code_change, on: :update
         after_validation :check_defaults
 
-        default_scope :order => [:position_value, :name]
+        default_scope { order([:position_value, :name]) }
 
-        scope :filter_archived_except_selected, lambda { |selected_id| {:conditions => ["archived_at is ? or id = ?", nil, selected_id.to_i]} }
+        scope :filter_archived_except_selected, lambda { |selected_id| where(["archived_at is ? or id = ?", nil, selected_id.to_i]) }
       end
 
       module ClassMethods
@@ -42,16 +42,46 @@ module Selections
         # Selection.priorities => returns [high, medium, low] instances (children)
 
         def method_missing lookup_code, *options
-          if (scope = where(:system_code => lookup_code.to_s)).exists?
+          if (scope = where(system_code: lookup_code.to_s)).exists?
             scope.first
-          elsif (scope = where(:system_code => lookup_code.to_s.singularize)).exists?
+          elsif (scope = where(system_code: lookup_code.to_s.singularize)).exists?
             scope.first.children
           else
             super
           end
         end
 
+        # label_to_id
+        #
+        # When using fixtures with the label same as the system_code, use this method to return the ID of the of the fixture
+        # and use this in Factories instead of using a lookup as it does not need a DB search.
+        #
+        # eg:
+        # Fixture File
+        #----------------------------------------
+        # priority_high:
+        #   name: Priorities
+        #   system_code: priority_high
+        #   parent: priority
+        #----------------------------------------
+        #
+        # in Factory
+        # ---------------------------------------
+        # priority: { Selection.priority_high }     <== Don't do this as it will need a DB lookup
+        #
+        # priority_id: { Selection.label_to_id(:priority_high) }    <== This will be much quicker
+        #
+        def label_to_id(label)
+          # ActiveRecord::FixtureSet replaces ActiveRecord::Fixtures in Rails 4
+          # This helps to ensure backwards compatibility
+          if ActiveRecord::VERSION::MAJOR >= 4
+            ActiveRecord::FixtureSet.identify(label)
+          else
+            ActiveRecord::Fixtures.identify(label)
+          end
+        end
       end
+
 
       def to_s #:nodoc:
         name.to_s
@@ -93,7 +123,7 @@ module Selections
       end
 
       def siblings_with_default_set #:nodoc:
-        self.parent.children.where(:is_default => true).where("id != ?", self.id.to_i).first
+        self.parent.children.where(is_default: true).where("id != ?", self.id.to_i).first
       end
 
       def archived #:nodoc:
@@ -111,6 +141,12 @@ module Selections
       def sub_children
         children.flat_map(&:children)
       end
+
+      # returns first default item in list e.g. Selection.priorities.default
+      def default
+        children.where(:is_default => true).order(:position_value).first
+      end
+
     end
 
     def selectable #:nodoc: