+++ CHANGELOG.md
@@ -1,15 +1,52 @@
-## 0.0.1
+## 1.0.1
+* lock test to rspec ~> 2.99
 
-* Initial release
+## 1.0.0
+* https://github.com/nigelr/selections/pull/14 - Update selections to no longer use hash rocket syntax.
+* https://github.com/nigelr/selections/pull/17 - Add rails 4 support, include appraisals gem and travis
+* https://github.com/nigelr/selections/pull/18 - Update readme to provide instructions on boot.rb 
+* https://github.com/nigelr/selections/pull/19 - Update generators for Rails 3 or 4
+* https://github.com/nigelr/selections/pull/20 - Return first default item in a group #13 
 
-## 0.1.1
+## 0.2.1
+* Check the existence of the selections table in `belongs_to_selection` so it is safe to use during rake tasks like db:reset.
 
-* Clean up of specs and readme
+## 0.1.14 (should of been 0.2.0)
+* https://github.com/nigelr/selections/issues/7 - add selection items as booleans method names to attached model
+
+## 0.1.13
+
+* https://github.com/nigelr/selections/issues/4 - selections-4 create selection Management scaffold generator
+## 0.1.12
+
+* https://github.com/nigelr/selections/issues/6 - #label_to_id feature to allow quick ID lookups within factories (speed enhancement)
+
+## 0.1.11
+
+* https://github.com/nigelr/selections/issues/2 - Added feature of radios for radio_buttons
+
+## 0.1.7
+
+* https://github.com/nigelr/selections/issues/5 - issue with forgetting selected item when page was invalid
+
+## 0.1.6
+
+* Fixed issue with options and html_options order
+
+## 0.1.3
+
+* Fixing issue with dependencies on Rubygems
 
 ## 0.1.2
 
 * Fixing issues with readme
 
-## 0.1.3
+## 0.1.1
 
-* Fixing issue with dependencies on Rubygems
+* Clean up of specs and readme
+
+## 0.0.1
+
+* Initial release
+
+